Update footer to use the renderable parts pattern
