+++ src/components/assets/Icons.js
@@ -107,6 +107,47 @@ export const ShareIcon = () =>
     </g>
   </SVGIcon>
 
+// From Footer
+export const PhoneIcon = () =>
+  <SVGIcon className="PhoneIcon">
+    <g>
+      <line x1="5" y1="14.1" x2="14.7" y2="14.1" />
+      <line x1="5" y1="5.9" x2="14.7" y2="5.9" />
+      <circle cx="9.9" cy="16.3" r="0.6" />
+      <circle cx="8.2" cy="3.7" r="0.6" />
+      <path d="M14.7,16.8c0,1-0.8,1.8-1.8,1.8H6.8c-1,0-1.8-0.8-1.8-1.8V3.2c0-1,0.8-1.8,1.8-1.8H13 c1,0,1.8,0.8,1.8,1.8V16.8z" />
+      <path d="M10.4,4.2C10,4.2,9.8,4,9.8,3.7s0.3-0.6,0.6-0.6h1.1c0.3,0,0.6,0.3,0.6,0.6s-0.3,0.6-0.6,0.6 H10.4z" />
+    </g>
+  </SVGIcon>
+
+// From Footer
+export const ChevronIcon = () =>
+  <SVGIcon className="ChevronIcon">
+    <g>
+      <polyline points="5,12.2 10.2,7 15.5,12.2" />
+    </g>
+  </SVGIcon>
+
+// From Footer
+export const ListIcon = () =>
+  <SVGIcon className="ListIcon">
+    <g>
+      <path d="M6.2,8.8c-1.4,0-2.5-1.1-2.5-2.5s1.1-2.5,2.5-2.5h7.5c1.4,0,2.5,1.1,2.5,2.5s-1.1,2.5-2.5,2.5H6.2z" />
+      <path d="M6.2,16.2c-1.4,0-2.5-1.1-2.5-2.5s1.1-2.5,2.5-2.5h7.5c1.4,0,2.5,1.1,2.5,2.5s-1.1,2.5-2.5,2.5H6.2z" />
+    </g>
+  </SVGIcon>
+
+// From Footer
+export const GridIcon = () =>
+  <SVGIcon className="GridIcon">
+    <g>
+      <circle cx="6.2" cy="6.2" r="2.5" />
+      <circle cx="13.8" cy="6.2" r="2.5" />
+      <circle cx="6.2" cy="13.8" r="2.5" />
+      <circle cx="13.8" cy="13.8" r="2.5" />
+    </g>
+  </SVGIcon>
+
 // -------------------------------------
 // SVG social icons
