--- src/components/footer/FooterLabel.css
@@ -1,8 +1,0 @@
-@import "../../lib/styles/properties.css";
-
-.FooterLabel {
-  display: inline-block;
-  font-size: 0.75rem;
-  vertical-align: middle;
-}
-
