--- src/components/footer/FooterLink.css
@@ -1,31 +1,0 @@
-@import "../../lib/styles/properties.css";
-
-.FooterLink {
-  display: inline-block;
-  font-size: 0.75rem;
-  vertical-align: middle;
-  border-bottom: 0;
-}
-
-.FooterLink > span {
-  display: inline-block;
-  overflow: hidden;
-  white-space: nowrap;
-  vertical-align: middle;
-  transition: width var(--speed) var(--ease);
-}
-
-.FooterLink > .SVGIcon + span {
-  width: 0;
-  margin-left: 5px;
-}
-
-.no-touch .FooterLink:hover {
-  color: #000;
-}
-
-/* // Only setup for `Get the app` link */
-.no-touch .FooterLink:hover > .SVGIcon + span { /* stylelint-disable-line */
-  width: 90px;
-}
-
