--- src/components/footer/FooterLink.js
@@ -1,24 +1,0 @@
-import React, { PropTypes } from 'react'
-import classNames from 'classnames'
-
-// This is alway an external link.
-export const FooterLink = ({ className, href, icon, label }) =>
-  <a
-    className={classNames(className, 'FooterLink')}
-    href={href}
-    rel="noopener noreferrer"
-    target="_blank"
-  >
-    {icon}
-    <span>{label}</span>
-  </a>
-
-FooterLink.propTypes = {
-  className: PropTypes.string,
-  href: PropTypes.string,
-  icon: PropTypes.node,
-  label: PropTypes.string,
-}
-
-export default FooterLink
-
