--- src/components/footer/FooterTool.css
@@ -1,39 +1,0 @@
-@import "../../lib/styles/properties.css";
-
-.FooterTool {
-  font-size: 0.75rem;
-  text-align: left;
-}
-
-.FooterTool > span {
-  display: inline-block;
-  width: 0;
-  margin-left: 5px;
-  overflow: hidden;
-  white-space: nowrap;
-  vertical-align: middle;
-  transition: width var(--speed) var(--ease), color var(--speed);
-}
-
-.FooterTool.LayoutTool > .SVGIcon {
-  margin-top: -2px;
-}
-
-.no-touch .FooterTool:hover {
-  color: #000;
-}
-
-.no-touch .FooterTool.TopTool:hover > span { /* stylelint-disable-line */
-  width: 30px;
-}
-
-.no-touch .FooterTool.LayoutTool:hover > span { /* stylelint-disable-line */
-  width: 66px;
-}
-
-@media (--break-2-max) {
-  .FooterTool.LayoutTool {
-    display: none;
-  }
-}
-
