--- src/components/footer/FooterTool.js
@@ -1,18 +1,0 @@
-import React, { PropTypes } from 'react'
-import classNames from 'classnames'
-
-export const FooterTool = ({ className, icon, label, onClick }) =>
-  <button className={classNames(className, 'FooterTool')} onClick={onClick} >
-    {icon}
-    <span>{label}</span>
-  </button>
-
-FooterTool.propTypes = {
-  className: PropTypes.string,
-  icon: PropTypes.node,
-  label: PropTypes.string,
-  onClick: PropTypes.func,
-}
-
-export default FooterTool
-
