+++ src/containers/FooterContainer.js
@@ -5,7 +5,7 @@ import { selectIsGridMode, selectIsLayoutToolHidden } from '../selectors/gui'
 import { selectStreamType } from '../selectors/stream'
 import { scrollTo } from '../lib/jello'
 import { LOAD_NEXT_CONTENT_REQUEST, SET_LAYOUT_MODE } from '../constants/action_types'
-import { Footer } from '../components/footer/Footer'
+import { Footer } from '../components/footer/FooterRenderables'
 
 function mapStateToProps(state) {
   const streamType = selectStreamType(state)
